Updating things. Camera temp disabled btw
There are a ton of changes. I have temporarily dissabled the camera :) no worries.
It'll soon be back up with new running code. The blueprint is written, will try and fix it entirely. The deathblock is already fixed. I was also thinking about adding permanent bodies on the place of the dead. actually, I would like to add guts and a skull :P that then permanently stay.